Add failing tests of prn writer
